Update bot for recent changes to Final Fantasy Randomizer (FF1R).
FF1R recently changed how we release so that each version is deployed
as a separate site.  As a result, players need to provide a full URL
string instead of just the flags so it generates a link to the correct
version.

Because this breaks the existing ff1r !flags command, I have updated
it and renamed it to !ff1url.  The documentation is also updated.

I also removed two legacy bot commands for a past tournament that are
no longer relevant (and also don't work).